refactor into components, openAI into content-script
