+++ src/content-script/index.js
@@ -1,13 +1,19 @@
 import { parseLinkedInChat } from "../utils/chatParser.js";
+import { callGPT4 } from "./openai.js";
 
 // Listen for messages from background script
 chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
   if (request.type === "PARSE_CHAT") {
     console.log("PARSE_CHAT");
     const messages = parseLinkedInChat();
-    sendResponse({ messages });
-    return true;
+
+    if (messages) {
+      callGPT4(messages).then((gptResponse) => sendResponse({ gptResponse }));
+    } else {
+      console.log("Message is empty");
+    }
   }
+  return true;
 });
 
 // Utility function to send parsed messages
