+++ src/content-script/openai.js
@@ -1,23 +1,13 @@
 import { PROFILE_MEMORY_KEY, OPENAI_API_KEY } from "../utils/constants";
 
 export async function callGPT4(messages) {
-  let profileData;
   try {
-    profileData = await new Promise((resolve, reject) => {
-      chrome.storage.local.get(PROFILE_MEMORY_KEY, function(result) {
-        if (chrome.runtime.lastError) {
-          reject(chrome.runtime.lastError);
-        } else {
-          resolve(result[PROFILE_MEMORY_KEY] || ""); // Default to empty string if not found
-        }
-      });
-    });
-  } catch (error) {
-    console.error("Error retrieving profile data:", error);
-    profileData = "";
-  }
-  try {
-    console.log("CALLED in content script");
+    const storageData = await chrome.storage.local.get(PROFILE_MEMORY_KEY);
+    const profileData = storageData?.[PROFILE_MEMORY_KEY] || "";
+    const content =
+      "You are a helpful assistant responding to LinkedIn messages on behalf of a the person receiving a certain LinkedIn message." +
+      `Answer any messages concisely, politely, and in a friendly tone. Use this profile data to help formulate your answers: ${profileData}`;
+
     const response = await fetch("https://api.openai.com/v1/chat/completions", {
       method: "POST",
       headers: {
@@ -29,9 +19,7 @@ export async function callGPT4(messages) {
         messages: [
           {
             role: "system",
-            content:
-              "You are a helpful assistant responding to LinkedIn messages on behalf of a the person receiving a certain LinkedIn message." +
-              `Answer any messages concisely, politely, and in a friendly tone. Use this profile data to help formulate your answers: ${profileData}`,
+            content,
           },
           {
             role: "user",
